third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/anonymous-iframe/anonymous-iframe-popup.tentative.https-expected.txt

This is a testharness.js-based test.
FAIL Cross-origin popup from anonymous iframe assert_equals: expected null but got object "[object Window]"
FAIL Same-origin popup from anonymous iframe assert_equals: expected null but got object "[object Window]"
Harness: the test ran to completion.
